backend/requirements.txt
FastAPI
uvicorn
psycopg2-binary
pydantic
pydantic-settings
pydantic[email]
sqlalchemy
python-multipart
transformers
torch
pdf2image
PyPDF2
Pillow
aiofiles
bcrypt
python-jose[cryptography]
passlib[bcrypt]
requests
python-dotenv